Improved module-wise variable usage
