+++ lazyqml/Factories/Models/QNNTorch.py
@@ -7,8 +7,7 @@ from lazyqml.Interfaces.iAnsatz import Ansatz
 from lazyqml.Interfaces.iCircuit import Circuit
 from lazyqml.Factories.Circuits.fCircuits import *
 from lazyqml.Global.globalEnums import Backend
-from lazyqml.Utils.Utils import printer
-import lazyqml
+from lazyqml.Utils import printer, get_simulation_type, get_max_bond_dim
 
 import warnings
 
@@ -26,16 +25,16 @@ class QNNTorch(Model):
         self.batch_size = batch_size
         self.backend = backend
 
-        if lazyqml.get_simulation_type() == "tensor":
+        if get_simulation_type() == "tensor":
             if backend != Backend.lightningTensor:
                 device_kwargs = {
-                    "max_bond_dim": lazyqml._max_bond_dim,
+                    "max_bond_dim": get_max_bond_dim(),
                     "cutoff": np.finfo(np.complex128).eps,
                     "contract": "auto-mps",
                 }
             else:
                 device_kwargs = {
-                    "max_bond_dim": lazyqml._max_bond_dim,
+                    "max_bond_dim": get_max_bond_dim(),
                     "cutoff": 1e-10,
                     "cutoff_mode": "abs",
                 }
@@ -71,7 +70,7 @@ class QNNTorch(Model):
         
 
         # Define the quantum circuit as a PennyLane qnode
-        @qml.qnode(self.deviceQ, interface='torch', diff_method='adjoint' if lazyqml.get_simulation_type() == "statevector" else 'best')
+        @qml.qnode(self.deviceQ, interface='torch', diff_method='adjoint' if get_simulation_type() == "statevector" else 'best')
         def circuit(x, theta):
             
             embedding.getCircuit()(x, wires=range(self.nqubits))
