+++ lazyqml/Utils/__init__.py
@@ -1,0 +1,2 @@
+from .Utils import get_max_bond_dim, get_simulation_type, set_max_bond_dim, set_simulation_type
+from .Utils import printer
