+++ lazyqml/_config.py
@@ -1,0 +1,2 @@
+_simulation = "statevector"
+_max_bond_dim = 64
